--- apps/server/src/modules/tldraw/index.ts
@@ -1,5 +1,0 @@
-export * from './tldraw.module';
-export * from './tldraw-test.module';
-export * from './tldraw-console.module';
-export * from './tldraw-ws.module';
-export * from './tldraw-ws-test.module';
